requirements.txt
flask
flask-sqlalchemy
flask-socketio
flask-login
python-dotenv
reportlab
pandas
openpyxl
flask-migrate
werkzeug
fastapi
uvicorn
pydantic
opencv-python-headless
